requirements.txt
flask==2.3.3
flask-cors==4.0.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
spotipy==2.23.0
python-dotenv==1.0.0
werkzeug==2.3.7